src/pyproject.toml
[tool.poetry]
name = "doogle"
version = "0.1.0"
description = ""
authors = ["alvarogf97 <ankora45@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
django-redis = "^5.2.0"
celery = "^5.2.6"
django-extensions = "^3.1.5"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
django-timezone-field = "^5.0"
python-crontab = "^2.6.0"
django-celery-beat = {git = "https://github.com/celery/django-celery-beat"}
tzdata = "^2022.1"
ipython = "^8.2.0"
django-localflavor = "^3.1"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"
django-polymorphic = "^3.1.0"
django-phonenumber-field = "^6.1.0"
phonenumberslite = "^8.12.46"
django-localized-fields = "^6.6"
Pillow = "^9.1.0"

[tool.poetry.dev-dependencies]
factory-boy = "^3.2.1"
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "doogle.settings"

[tool.coverage.paths]
source = ["./", "/app/"]

[tool.coverage.run]
omit = ["./doogle/*.py", "*tests/*", "**/admin/*.py", "*/admin.py", "**/migrations/*.py", "**/manage.py"]
